Implemented repositories

+++ IMT-Backend/IMT-Backend.Domain/Entities/Income.cs
@@ -5,7 +5,7 @@ namespace IMT_Backend.Domain.Entities
 {
     public class Income : BaseEntity
     {
-        public float Quantity { get; set; }
+        public double Quantity { get; set; }
         public IncomeCategory Category { get; set; }
         public string UserId { get; set; }
         public User User { get; set; }
