+++ IMT-Backend/IMT-Backend.Domain/Entities/User.cs
@@ -4,7 +4,7 @@
     {
         public string Id { get; set; }
         public string Username { get; set; }
-        public float Sum { get; set; }
+        public double Sum { get; set; }
         public List<Income> Incomes { get; set; }
         public List<Expense> Expenses { get; set; }
     }
